+++ yash-syntax/src/parser/lex/raw_param.rs
@@ -18,6 +18,8 @@
 
 use super::core::Lexer;
 use crate::parser::core::Result;
+use crate::syntax::Param;
+use crate::syntax::ParamType;
 use crate::syntax::SpecialParam;
 use crate::syntax::TextUnit;
 
@@ -37,13 +39,13 @@ pub const fn is_special_parameter_char(c: char) -> bool {
     SpecialParam::from_char(c).is_some()
 }
 
-/// Tests if a character is a valid single-character raw parameter name.
+/// Tests if a character is a valid single-character raw parameter.
 ///
-/// If this function returns true, the character is a valid parameter name for a
-/// raw parameter expansion, but the next character is never treated as part of
-/// the name.
+/// If this function returns true, the character is a valid parameter for a raw
+/// parameter expansion, but the next character is never treated as part of the
+/// parameter.
 ///
-/// This function returns true for ASCII digits and special parameter names.
+/// This function returns true for ASCII digits and special parameters.
 pub const fn is_single_char_name(c: char) -> bool {
     c.is_ascii_digit() || is_special_parameter_char(c)
 }
@@ -60,19 +62,29 @@ impl Lexer<'_> {
     /// used to construct the result, but this function does not check if it
     /// actually points to the `$`.
     pub async fn raw_param(&mut self, start_index: usize) -> Result<Option<TextUnit>> {
-        let name = if let Some(c) = self.consume_char_if(is_single_char_name).await? {
-            c.value.to_string()
+        let param = if let Some(c) = self.consume_char_if(is_special_parameter_char).await? {
+            Param {
+                id: c.value.to_string(),
+                r#type: SpecialParam::from_char(c.value).unwrap().into(),
+            }
+        } else if let Some(c) = self.consume_char_if(|c| c.is_ascii_digit()).await? {
+            Param {
+                id: c.value.to_string(),
+                r#type: ParamType::Positional(c.value.to_digit(10).unwrap() as usize),
+            }
         } else if let Some(c) = self.consume_char_if(is_portable_name_char).await? {
             let mut name = c.value.to_string();
             while let Some(c) = self.consume_char_if(is_portable_name_char).await? {
                 name.push(c.value);
             }
-            name
+            Param::variable(name)
         } else {
             return Ok(None);
         };
+
         let location = self.location_range(start_index..self.index());
-        Ok(Some(TextUnit::RawParam { name, location }))
+
+        Ok(Some(TextUnit::RawParam { param, location }))
     }
 }
 
@@ -80,6 +92,7 @@ impl Lexer<'_> {
 mod tests {
     use super::*;
     use crate::source::Source;
+    use crate::syntax::Param;
     use assert_matches::assert_matches;
     use futures_util::FutureExt;
 
@@ -90,8 +103,8 @@ mod tests {
         lexer.consume_char();
 
         let result = lexer.raw_param(0).now_or_never().unwrap().unwrap().unwrap();
-        assert_matches!(result, TextUnit::RawParam { name, location } => {
-            assert_eq!(name, "@");
+        assert_matches!(result, TextUnit::RawParam { param, location } => {
+            assert_eq!(param, Param::from(SpecialParam::At));
             assert_eq!(*location.code.value.borrow(), "$@;");
             assert_eq!(location.code.start_line_number.get(), 1);
             assert_eq!(*location.code.source, Source::Unknown);
@@ -108,8 +121,8 @@ mod tests {
         lexer.consume_char();
 
         let result = lexer.raw_param(0).now_or_never().unwrap().unwrap().unwrap();
-        assert_matches!(result, TextUnit::RawParam { name, location } => {
-            assert_eq!(name, "1");
+        assert_matches!(result, TextUnit::RawParam { param, location } => {
+            assert_eq!(param, Param::from(1));
             assert_eq!(*location.code.value.borrow(), "$12");
             assert_eq!(location.code.start_line_number.get(), 1);
             assert_eq!(*location.code.source, Source::Unknown);
@@ -126,8 +139,8 @@ mod tests {
         lexer.consume_char();
 
         let result = lexer.raw_param(0).now_or_never().unwrap().unwrap().unwrap();
-        assert_matches!(result, TextUnit::RawParam { name, location } => {
-            assert_eq!(name, "az_AZ_019");
+        assert_matches!(result, TextUnit::RawParam { param, location } => {
+            assert_eq!(param, Param::variable("az_AZ_019"));
             assert_eq!(*location.code.value.borrow(), "$az_AZ_019<");
             assert_eq!(location.code.start_line_number.get(), 1);
             assert_eq!(*location.code.source, Source::Unknown);
@@ -144,8 +157,8 @@ mod tests {
         lexer.consume_char();
 
         let result = lexer.raw_param(0).now_or_never().unwrap().unwrap().unwrap();
-        assert_matches!(result, TextUnit::RawParam { name, location } => {
-            assert_eq!(name, "abc");
+        assert_matches!(result, TextUnit::RawParam { param, location } => {
+            assert_eq!(param, Param::variable("abc"));
             assert_eq!(*location.code.value.borrow(), "$a\\\n\\\nb\\\n\\\nc\\\n>");
             assert_eq!(location.code.start_line_number.get(), 1);
             assert_eq!(*location.code.source, Source::Unknown);